@@ -3,11 +3,11 @@
   "description": "Library that handles fractions as nominator/denominator",
   "license": "MIT",
   "require": {
-    "php": "7.*",
+    "php": "8.3.*",
     "ext-gmp": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "7.2.7"
+    "phpunit/phpunit": "12.0.7"
   },
   "autoload": {
     "psr-4": {
@@ -21,5 +21,10 @@
   },
   "config": {
     "optimize-autoloader": true
+  },
+  "scripts": {
+    "test": [
+      "./vendor/bin/phpunit --testdox ./tests/"
+    ]
   }
 }
