@@ -8,6 +8,7 @@ use pbaczek\fraction\Exceptions\NegativeDenominatorException;
 use pbaczek\fraction\Exceptions\ZeroDenominatorException;
 use pbaczek\fraction\Fraction;
 use pbaczek\fraction\MFraction;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -16,8 +17,7 @@ use PHPUnit\Framework\TestCase;
  */
 class MFractionTest extends TestCase
 {
-    /** @var MFraction $mFraction */
-    private $mFraction;
+    private MFraction $mFraction;
 
     /**
      * @inheritDoc
@@ -188,7 +188,7 @@ class MFractionTest extends TestCase
      * Defined functions for
      * @return array
      */
-    public function definedFunctions(): array
+    public static function definedFunctions(): array
     {
         return [
             [
@@ -206,12 +206,7 @@ class MFractionTest extends TestCase
         ];
     }
 
-    /**
-     * Test that only same type objects can be added
-     * @dataProvider definedFunctions
-     * @param string $function
-     * @return void
-     */
+    #[DataProvider('definedFunctions')]
     public function testAllMathFunctionsWorkOnlyOnSameObject(string $function): void
     {
         $this->expectException(InvalidArgumentException::class);
